+++ .github/workflows/cypress-tests.yml
@@ -1,0 +1,31 @@
+name: Cypress Tests
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Use Node.js
+        uses: actions/setup-node@v2
+        with:
+          node-version: '20.9.0'
+
+      - name: Install dependencies
+        run: npm install
+
+      - name: Run Cypress tests
+        run: npm run cy:run -- --spec "cypress/e2e/AddingCart.js"
+
+      - name: Archive test results and artifacts
+        uses: actions/upload-artifact@v2
+        with:
+          name: cypress-results
+          path: cypress/artifacts/
